src/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
celery>=5.1.2
redis>=3.5.3
psycopg2-binary>=2.9.1
numpy>=1.21.1
torch>=1.9.0
pydantic>=1.8.2
requests>=2.26.0